Add a possibility to write when no root permission
When you open and edit a file with vim without root permission, you can not
write the file without closing without saving and editing the file another time.
The ':Wroot' will allow you to save the current file with the root permission.
WARNING: The file is written in the stdin then piped to the same filename with
root permission; no undo will be available with this solution and the file
should be reload at the end of the command.